+++ pure_extensions/lib/path_extensions.dart
@@ -21,14 +21,15 @@ extension StringPathExtension on String {
 
   bool get isRootRelativePath => pathExtensionContext.isRootRelative(this);
 
-  String joinPaths(
-      [String part1,
-      String part2,
-      String part3,
-      String part4,
-      String part5,
-      String part6,
-      String part7]) {
+  String joinPaths([
+    String part1,
+    String part2,
+    String part3,
+    String part4,
+    String part5,
+    String part6,
+    String part7,
+  ]) {
     return pathExtensionContext.join(this, part1, part2, part3, part4, part5, part6, part7);
   }